config: mock media query globally for unit tests
