+++ src/App.module.css
@@ -99,19 +99,6 @@
   cursor: pointer;
 }
 
-.nav {
-  height: 100%;
-  background: var(--colors-bgPrimary);
-  display: flex;
-  flex-direction: row;
-  justify-content: space-between;
-  align-items: center;
-  padding-left: var(--spacing-sm);
-  padding-right: var(--spacing-sm);
-  padding-top: var(--spacing-sm);
-}
-
-
 @media (min-width: 768px) {
   .sidebar {
     transform: translateX(0);
