+++ src/NavBar.module.css
@@ -7,7 +7,6 @@
   align-items: center;
   padding-left: var(--spacing-sm);
   padding-right: var(--spacing-sm);
-  padding-top: var(--spacing-md);
 }
 
 .leftSection {
