+++ src/setupTests.ts
@@ -1,6 +1,10 @@
 import '@testing-library/jest-dom';
-import { afterEach } from 'vitest';
+import { afterEach, vi } from 'vitest';
 import { cleanup } from '@testing-library/react';
+
+vi.mock('./theme/useIsMobileLayout', () => ({
+  useIsMobileLayout: vi.fn().mockReturnValue(false),
+}));
 
 afterEach(() => {
   cleanup();
