+++ src/theme/useIsMobileLayout.ts
@@ -1,21 +1,3 @@
-import { useEffect, useState } from 'react';
-
-const useMediaQuery = (query: string) => {
-  const [matches, setMatches] = useState<boolean>(
-    window.matchMedia(query).matches,
-  );
-
-  useEffect(() => {
-    const matchQuery = window.matchMedia(query);
-    const onChange = (e: MediaQueryListEvent) => {
-      setMatches(e.matches);
-    };
-
-    matchQuery.addEventListener('change', onChange);
-    return () => matchQuery.removeEventListener('change', onChange);
-  }, [query]);
-
-  return matches;
-};
+import { useMediaQuery } from './useMediaQuery';
 
 export const useIsMobileLayout = () => useMediaQuery('(max-width: 768px)');
